Split SvgClosePathSegment
